+++ ZAPDUtils/Utils/Directory.h
@@ -1,8 +1,6 @@
 #pragma once
 
-#include <iostream>
 #include <string>
-#include <vector>
 
 #if __has_include(<filesystem>)
 #include <filesystem>
@@ -17,23 +15,7 @@ namespace fs = std::experimental::filesystem;
 class Directory
 {
 public:
-	static std::string GetCurrentDirectory() { return fs::current_path().u8string(); }
-
-	static bool Exists(const fs::path& path) { return fs::exists(path); }
-
-	static void CreateDirectory(const std::string& path)
-	{
-		std::string curPath = "";
-		std::vector<std::string> split = StringHelper::Split(path, "/");
-
-		for (std::string s : split)
-		{
-			curPath += s + "/";
-
-			if (!Exists(curPath))
-				fs::create_directory(curPath);
-		}
-
-		// fs::create_directory(path);
-	}
+	static std::string GetCurrentDirectory();
+	static bool Exists(const fs::path& path);
+	static void CreateDirectory(const std::string& path);
 };
