+++ ZAPDUtils/Utils/Stream.cpp
@@ -1,0 +1,8 @@
+#include "Stream.h"
+
+#include <stdint.h>
+
+uint64_t Stream::GetBaseAddress()
+{
+	return baseAddress;
+}
