+++ ZAPDUtils/Vec3f.h
@@ -6,16 +6,6 @@ struct Vec3f
 {
 	float x, y, z;
 
-	Vec3f()
-	{
-		x = 0;
-		y = 0;
-		z = 0;
-	};
-	Vec3f(float nX, float nY, float nZ)
-	{
-		x = nX;
-		y = nY;
-		z = nZ;
-	};
+	Vec3f() : x(0), y(0), z(0) {}
+	Vec3f(float nX, float nY, float nZ) : x(nX), y(nY), z(nZ) {}
 };
